+++ api/fe.js
@@ -1,0 +1,7 @@
+const { generateScript } = require('../forward_engineering/generateScript');
+const { generateContainerScript } = require('../forward_engineering/generateContainerScript');
+
+module.exports = {
+	generateScript,
+	generateContainerScript,
+};
