+++ forward_engineering/helpers/getContainerThroughputProps.js
@@ -1,0 +1,13 @@
+const getContainerThroughputProps = containerData => {
+	if (containerData?.capacityMode === 'Serverless') {
+		return {};
+	}
+	if (containerData?.autopilot) {
+		return { maxThroughput: containerData.throughput || 4000 };
+	}
+	return { throughput: containerData?.throughput || 400 };
+};
+
+module.exports = {
+	getContainerThroughputProps,
+};
