+++ forward_engineering/helpers/updateSample.js
@@ -1,0 +1,16 @@
+const updateSample = (sample, containerData, entityData) => {
+	const docType = containerData?.docTypeName;
+
+	if (!docType) {
+		return sample;
+	}
+
+	return {
+		...sample,
+		[docType]: entityData.code || entityData.collectionName,
+	};
+};
+
+module.exports = {
+	updateSample,
+};
